[ForceField] Add direction option in RestShapeSpringsFF
The idea of this commit is to add the ability for a user to select in
which directions the RestShapeSpringsFF will be effective during
simulation. For this we add a new data field containing 6 booleans: one
for each direction of effort (3 directions of translation, 3 of
rotation). For each entry, "true" indicates that the direction will be
taken into account, and "false" that it won't. The default value of the
parameter is set to [1 1 1 1 1 1], so that all directions are taken into
account if the parameter is not specifically provided by the user. This
way, we expect the component not to be breaking.

/!\ At the moment, we only implemented the restriction for translations,
but not for rotations. Concretely, changing one of the last 3 entries of
the data field to "false" doesn't have any effect on the simulation. The
implementation for rotations should be the object of further commits.